Make NodeToCode static.
