+++ src/DynamoCore/DynamoCore.csproj
@@ -118,7 +118,7 @@ limitations under the License.
     <Compile Include="Core\DynamoMigrator.cs" />
     <Compile Include="Exceptions\LibraryLoadFailedException.cs" />
     <Compile Include="Graph\Workspaces\Layout.cs" />
-    <Compile Include="Graph\Workspaces\NodeToCodeReconnection.cs" />
+    <Compile Include="Graph\Workspaces\NodesToCode.cs" />
     <Compile Include="Graph\Workspaces\Presets.cs" />
     <Compile Include="Graph\Workspaces\Serialiaztion\SerializationConverters.cs" />
     <Compile Include="Graph\Workspaces\Serialiaztion\Workspaces.cs" />
